Updated Dockerfile and Readme.md
Dockerfile was updated to run Node-RED as the user `node-red` as well as
expose the `/data` directory as a Docker volume. This directory is the new storage
location for Node-RED config and flow data. I have also reverted the default behavior to
expose Node-RED on it's default tcp port of `1880`